+++ src/core/init/init_clock.h
@@ -1,7 +1,7 @@
 #pragma once
 
 typedef struct ClockStateInit {
-    double tick_interval;    
+    int tick_interval;    
 } ClockInit;
 
 extern const ClockInit c_clock_init;
